Phonegap 3.0 compatability (new plugin format and ARC compatible)
I've made the plugin compatible with phonegap 3.0 BUT I also tweaked
the plugin to work with files on iOS the way I was storing them.  Go
ahead and try the plugin and see what if it works for you.  You can
also just compare the CDVVideo.m file with the original to see what I
changed and change it back.

Discaimer: I'm not a seasoned objective c programer so use this at your
own risk and please correct me where I made mistakes.